更改页面结构

+++ platform-shop/src/main/java/com/platform/controller/GoodsController.java
@@ -48,7 +48,7 @@ public class GoodsController {
      */
     @RequestMapping("/info/{id}")
     @RequiresPermissions("goods:info")
-    public R info(@PathVariable("id") Integer id) {
+    public R info(@PathVariable("id") Long id) {
         GoodsEntity goods = goodsService.queryObject(id);
 
         return R.ok().put("goods", goods);
@@ -81,7 +81,7 @@ public class GoodsController {
      */
     @RequestMapping("/delete")
     @RequiresPermissions("goods:delete")
-    public R delete(@RequestBody Integer[] ids) {
+    public R delete(@RequestBody Long[] ids) {
         goodsService.deleteBatch(ids);
 
         return R.ok();
@@ -125,7 +125,7 @@ public class GoodsController {
      */
     @RequestMapping("/back")
     @RequiresPermissions("goods:back")
-    public R back(@RequestBody Integer[] ids) {
+    public R back(@RequestBody Long[] ids) {
         goodsService.back(ids);
 
         return R.ok();
@@ -146,7 +146,7 @@ public class GoodsController {
      * 上架
      */
     @RequestMapping("/enSale")
-    public R enSale(@RequestBody Integer id) {
+    public R enSale(@RequestBody Long id) {
         goodsService.enSale(id);
 
         return R.ok();
@@ -156,7 +156,7 @@ public class GoodsController {
      * 下架
      */
     @RequestMapping("/unSale")
-    public R unSale(@RequestBody Integer id) {
+    public R unSale(@RequestBody Long id) {
         goodsService.unSale(id);
 
         return R.ok();
